feat(plugin-file-manager): expose preview registry on client-v2 and consolidate shared previewer
- Move src/shared/previewer/filePreviewTypes into src/client-v2/previewer; v1 client
  now re-exports it via relative path. Drop duplicate src/shared/locale in favor of
  the existing common/constants NAMESPACE, removing src/shared entirely.
- Re-export filePreviewTypes / wrapWithModalPreviewer (and their types) from the
  client-v2 entry so file-previewer plugins can register custom previews under v2
  without importing the v1 client entry.
- fix: prefix placeholder icon paths with the runtime public path
  (window.__nocobase_public_path__) so they resolve when APP_PUBLIC_PATH !== '/'.
