+++ packages/plugins/@nocobase/plugin-file-manager/src/client-v2/index.tsx
@@ -36,3 +36,8 @@ export {
 export type { DefaultFieldProps } from './components/DefaultField';
 export type { PathFieldProps } from './components/PathField';
 export { CardUpload, UploadFieldModel } from './models/UploadFieldModel';
+
+// Preview registry consumed by file-previewer plugins (e.g. plugin-file-previewer-office)
+// to add custom preview handlers under v2 without going through the v1 `@nocobase/plugin-file-manager/client` entry.
+export { filePreviewTypes, wrapWithModalPreviewer } from './previewer/filePreviewTypes';
+export type { FilePreviewType, FilePreviewerProps } from './previewer/filePreviewTypes';
