+++ packages/plugins/@nocobase/plugin-file-manager/src/client-v2/locale.ts
@@ -8,7 +8,7 @@
  */
 
 import { tExpr as flowTExpr, useFlowEngine } from '@nocobase/flow-engine';
-import { NAMESPACE } from '../shared/locale';
+import { NAMESPACE } from '../common/constants';
 
 export function useT() {
   const engine = useFlowEngine();
