+++ packages/plugins/@nocobase/plugin-file-manager/src/client-v2/models/uploadFieldUtils.ts
@@ -7,7 +7,7 @@
  * For more information, please refer to: https://www.nocobase.com/agreement.
  */
 
-import { getPreviewThumbnailUrl } from '../../shared/previewer/filePreviewTypes';
+import { getPreviewThumbnailUrl } from '../previewer/filePreviewTypes';
 
 interface UploadFieldFile {
   uid?: string;
