+++ packages/plugins/@nocobase/plugin-file-manager/src/client/locale/index.ts
@@ -8,7 +8,7 @@
  */
 
 import { useTranslation } from 'react-i18next';
-import { NAMESPACE } from '../../shared/locale';
+import { NAMESPACE } from '../../common/constants';
 
 export { NAMESPACE };
