+++ packages/plugins/@nocobase/plugin-file-manager/src/client/previewer/filePreviewTypes.tsx
@@ -7,4 +7,4 @@
  * For more information, please refer to: https://www.nocobase.com/agreement.
  */
 
-export * from '../../shared/previewer/filePreviewTypes';
+export * from '../../client-v2/previewer/filePreviewTypes';
